+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 3.9.2 (Upcoming)
 
+Merged from 3.0.9.1:
+* Add support for geospatial shapes in bounding box search
+* Add support for geospatial shapes in distance search
+* Improve performance of needs before write calculation
+
 ## 3.9.1 (October 17, 2016)
 
 Merged from 3.8.1: